// webui/src/proto/byte/v/forge/waapp/v1/profile.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.8
//   protoc               v7.34.1
// source: byte/v/forge/waapp/v1/profile.proto

/* eslint-disable */
import type { Account } from "@byte-v-forge/common-ui/proto/byte/v/forge/contracts/account/v1/account";
import type { AuditStamp, PhoneTarget, RequestContext, WaError } from "./common";

export const protobufPackage = "byte.v.forge.waapp.v1";

export enum WAAccountStatus {
  WA_ACCOUNT_STATUS_UNSPECIFIED = "WA_ACCOUNT_STATUS_UNSPECIFIED",
  WA_ACCOUNT_STATUS_ACTIVE = "WA_ACCOUNT_STATUS_ACTIVE",
  WA_ACCOUNT_STATUS_PAUSED = "WA_ACCOUNT_STATUS_PAUSED",
  WA_ACCOUNT_STATUS_ARCHIVED = "WA_ACCOUNT_STATUS_ARCHIVED",
  WA_ACCOUNT_STATUS_PENDING_REGISTRATION = "WA_ACCOUNT_STATUS_PENDING_REGISTRATION",
  UNRECOGNIZED = "UNRECOGNIZED",
}

export enum ClientProfileStatus {
  CLIENT_PROFILE_STATUS_UNSPECIFIED = "CLIENT_PROFILE_STATUS_UNSPECIFIED",
  CLIENT_PROFILE_STATUS_PREPARING = "CLIENT_PROFILE_STATUS_PREPARING",
  CLIENT_PROFILE_STATUS_READY = "CLIENT_PROFILE_STATUS_READY",
  CLIENT_PROFILE_STATUS_REJECTED = "CLIENT_PROFILE_STATUS_REJECTED",
  CLIENT_PROFILE_STATUS_RETIRED = "CLIENT_PROFILE_STATUS_RETIRED",
  UNRECOGNIZED = "UNRECOGNIZED",
}

export enum KeyMaterialStatus {
  KEY_MATERIAL_STATUS_UNSPECIFIED = "KEY_MATERIAL_STATUS_UNSPECIFIED",
  KEY_MATERIAL_STATUS_PENDING = "KEY_MATERIAL_STATUS_PENDING",
  KEY_MATERIAL_STATUS_READY = "KEY_MATERIAL_STATUS_READY",
  KEY_MATERIAL_STATUS_INVALID = "KEY_MATERIAL_STATUS_INVALID",
  KEY_MATERIAL_STATUS_ROTATED = "KEY_MATERIAL_STATUS_ROTATED",
  UNRECOGNIZED = "UNRECOGNIZED",
}

export interface WAAccount {
  account: Account | undefined;
  workspace_id: string;
  phone: PhoneTarget | undefined;
}

export interface ClientProfile {
  client_profile_id: string;
  wa_account_id: string;
  protocol_profile_id: string;
  status: ClientProfileStatus;
  registration_key_state: KeyMaterialStatus;
  messaging_key_state: KeyMaterialStatus;
  last_used_at: string | undefined;
  audit: AuditStamp | undefined;
  last_error: WaError | undefined;
}

export interface CreateWAAccountRequest {
  context: RequestContext | undefined;
  phone: PhoneTarget | undefined;
}

export interface CreateWAAccountResponse {
  account: WAAccount | undefined;
  error: WaError | undefined;
}

export interface GetWAAccountRequest {
  context: RequestContext | undefined;
  wa_account_id: string;
}

export interface GetWAAccountResponse {
  account: WAAccount | undefined;
  error: WaError | undefined;
}

export interface ListWAAccountsRequest {
  context: RequestContext | undefined;
  limit: number;
  cursor: string;
}

export interface ListWAAccountsResponse {
  accounts: WAAccount[];
  error: WaError | undefined;
  next_cursor: string;
}

export interface PrepareClientProfileRequest {
  context: RequestContext | undefined;
  wa_account_id: string;
  protocol_profile_id: string;
}

export interface PrepareClientProfileResponse {
  client_profile: ClientProfile | undefined;
  error: WaError | undefined;
}

export interface GetClientProfileRequest {
  context: RequestContext | undefined;
  client_profile_id: string;
}

export interface GetClientProfileResponse {
  client_profile: ClientProfile | undefined;
  error: WaError | undefined;
}

export interface RetireClientProfileRequest {
  context: RequestContext | undefined;
  client_profile_id: string;
  reason: string;
}

export interface RetireClientProfileResponse {
  client_profile: ClientProfile | undefined;
  error: WaError | undefined;
}